// Frontend/src/app/CRUD/models/student.ts
export class Student {
    studentId: number;
    courseId: string;
    id_School: number;
    metabolic_age: number;
    average_of_fat: number;
    average_of_Hydration: number;
    muscle_weight: number;
    muscle_mass_level: number;
    bone_weight: number;
    kilocalories: number;
    registration_date: Date;
    BMI: number;
    weight: number;
    height: number;
    abdominal_perimeter: number;
    physical_activity_level: number;
    year: number;
    sex: string;
}